dom/simplepush/test/test_register.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1235050: SimplePush basics tests.

Any copyright is dedicated to the Public Domain.
http://creativecommons.org/licenses/publicdomain/

-->
<head>
  <title>Test for Bug 1235050</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta http-equiv="Content-type" content="text/html;charset=UTF-8">
</head>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1235050">Mozilla Bug 1235050</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>

<script class="testbody" type="text/javascript">

  function test_navigator() {
    ok(("push" in navigator), "push object is exposed");
    ok(("register" in navigator.push), "push object is exposed with register");
    ok(("unregister" in navigator.push), "push object is exposed with unregister");
  }

  function test_register(next) {
    var pushReq = navigator.push.register();
    pushReq.onsuccess = e => {
      ok(true, "push.register() triggered onsuccess");
      next();
    };
    pushReq.onerror = e => {
      ok(false, "push.register() triggered onerror");
      next();
    };
  }

  function runTest() {
    test_navigator();
    test_register(() => {
      SimpleTest.finish();
    });
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.addPermission("push", true, document);

  runTest();
</script>
</body>
</html>
